added the Image component
This part provides the Image lib, a collection of classes dealing with several
image formats and base drawing functionality. There is the class Base from
which every Image class should inherit from. This class is not done yet, but
already works very well. There is also the first image format PPM, which is
also work in progress, but will be the first finished one.