updated Targa image class
Changed the monochrome pixel data loaders to keep grey scale data, if the
image was saved with grey scale data. This is actually not part of the spec,
but by keeping monochrome pixel data in byte instead of bits, it is possible
and not really wrong. Also added floyd-steinberg and burkes algorithms to the
monochrome rle saver. Burkes is the active one, floyed-steinberg is disabled.